+++ share/src/main/resources/alfresco/site-webscripts/com/parashift/onlyoffice-edit.get.js
@@ -6,6 +6,7 @@
 var query = "?nodeRef=" + url.args.nodeRef;
 if (url.args.readonly) query += "&readonly=1";
 if (url.args.new) query += "&new=" + url.args.new;
+if (url.args.parentNodeRef) query+= "&parentNodeRef=" + url.args.parentNodeRef;
 
 pObj = eval('(' + remote.call("/parashift/onlyoffice/prepare" + query) + ')');
 model.onlyofficeUrl = pObj.onlyofficeUrl;
@@ -13,4 +14,7 @@ model.docTitle = pObj.config.document.title;
 model.config = JSON.stringify(pObj.config);
 model.demo = pObj.demo;
 model.documentType=pObj.documentType;
+model.share = pObj.share;
+model.user = pObj.user;
+model.favorite = pObj.favorite;
 model.historyObj = pObj.historyObj;